% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{constrained.boot}
\alias{constrained.boot}
\title{Constrained bootstrap function}
\usage{
constrained.boot(df.window, effortData)
}
\arguments{
\item{df.window}{Dataframe with relational data in the first two rows, and a time stamp in the third row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{effortData}{Dataframe with the date (as the 1st column) and total number of scans(as the second column)}
}
\description{
This function will make sure that the individuals from 'from' column can only appear no more than total number of unique scans. This assumes that data are collected in discrete sampling occurences where individuals can only be recorded once.
}
