% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{convergence.check.boot.graph}
\alias{convergence.check.boot.graph}
\title{Bootstrap convergence check for network level measures}
\usage{
convergence.check.boot.graph(data, windowsize = days(30),
  windowshift = days(1), directed = FALSE, measureFun = eigen_mean,
  boot.samples = 100, SRI = FALSE, probs = c(0.025, 0.975))
}
\arguments{
\item{data}{Dataframe with relational data in the first two rows, and a time stamp in the third row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{windowsize}{The size of each window in which to generate a network.}

\item{windowshift}{The amount of time to shift the window when generating networks.}

\item{directed}{Whether to consider the network as directed or not (TRUE/FALSE).}

\item{measureFun}{The measurment function to perform the bootstap on (should be at the node level).}

\item{boot.samples}{The number of bootstrapped samples to run (Default=100)}

\item{SRI}{Wether to use the simple ratio index (Default=FALSE)}

\item{probs}{The quantiles of the bootrap samples to return (Default=c(0.025,0.975)).}
}
\description{
This function will estimate the convergence of the chosen network measure using bootstrapped samples of the data.
}
