% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{convergence.check.value}
\alias{convergence.check.value}
\title{Convergence check}
\usage{
convergence.check.value(data, windowsize = days(30), windowshift = days(1),
  directed = FALSE, measureFun = strength, max.subsample.size = 30,
  SRI = FALSE, n.boot = 100, probs = c(0.025, 0.975))
}
\arguments{
\item{data}{Dataframe with relational data in the first two rows, and a time stamp in the third row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{windowsize}{The size of each window in which to generate a network.}

\item{windowshift}{The amount of time to shift the window when generating networks.}

\item{directed}{Whether to consider the network as directed or not (TRUE/FALSE).}

\item{measureFun}{The function used to take network measurements.}

\item{max.subsample.size}{The maximum number of samples to remove when recalculating the network metric.}

\item{SRI}{Wether to use the simple ratio index or not (Default=FALSE).}

\item{n.boot}{Number of times to repeate the subsample procedure (i.e., bootstrap estimate of slope)}

\item{probs}{The probablility interval to calculate the confidence intervals.}
}
\description{
This function will estimate the convergence of the chosen network measure using random subsets of the data.
}
