% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{convergence.check.var}
\alias{convergence.check.var}
\title{Variance by window size check}
\usage{
convergence.check.var(data, windowsize_min = days(10),
  windowsize_max = days(40), by = days(1), windowshift = days(1),
  directed = FALSE, measureFun = igraph::edge_density, SRI = FALSE)
}
\arguments{
\item{data}{Dataframe with relational data in the first two rows, and a time stamp in the third row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{windowsize_min}{The min size of windowsize to test.}

\item{windowsize_max}{The max size of windowsize to test.}

\item{by}{The resolution at which to test window sizes between the min and the max window sizes}

\item{windowshift}{The amount of time to shift the window when generating networks.}

\item{directed}{Whether to consider the network as directed or not (TRUE/FALSE).}

\item{measureFun}{The measurment function to perform the bootstap on (Default: density).}

\item{SRI}{Wether to use the simple ratio index (Default=FALSE)}
}
\description{
This function will estimate the variance in the time series based on the window size.
}
