% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurment_functions_GRAPH.R
\name{cosine_between_graphs}
\alias{cosine_between_graphs}
\title{Estimate consine similarity between graphs}
\usage{
cosine_between_graphs(
  graph1,
  graph2,
  directed = FALSE,
  considerZeros = TRUE,
  center = FALSE
)
}
\arguments{
\item{graph1}{first graph (igraph)}

\item{graph2}{second graph (igraph)}

\item{directed}{Whether the graph is directed or not (default = FALSE)}

\item{considerZeros}{Whether to treat edges present in one graph, but not the other, as zero (TRUE), or consider only those edges present in both graphs (FALSE).}

\item{center}{Whether to center the resulting vectors before measuring cosine similarity.}
}
\description{
This function will calculate the cosine similarity between two graphs.
}
\examples{

#two random graphs
library(igraph)
graph1 <-erdos.renyi.game(n=10,p=0.1)
graph2 <-erdos.renyi.game(n=10,p=0.1)
cosine_between_graphs(graph1,graph2)

}
