% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurment_functions_NODES.R
\name{cosine_between_nodes}
\alias{cosine_between_nodes}
\title{Estimate consine similarity between nodes in a graph}
\usage{
cosine_between_nodes(graph1, graph2, directed = FALSE, mode = "out")
}
\arguments{
\item{graph1}{first graph (igraph)}

\item{graph2}{second graph (igraph)}

\item{directed}{Boolean indicating if the graph is directed or not.}

\item{mode}{For directed graphs this allows for in, out, or total cosine similarity to be calculated for each node. This argument is ignored for undirected graphs.}
}
\description{
This function will calculate the cosine similarity between nodes at two time periods.
}
\examples{

#two random graphs
library(igraph)
graph1 <-erdos.renyi.game(n=10,p=0.1)
graph2 <-erdos.renyi.game(n=10,p=0.1)
cosine_between_nodes(graph1,graph2)


}
