% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{create.a.network.SRI}
\alias{create.a.network.SRI}
\title{Create a network from an edge list using the SRI.}
\usage{
create.a.network.SRI(events)
}
\arguments{
\item{events}{dataframe containing all events.}
}
\description{
This function will generate an SRI network from an event dataframe
}
