% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{create.an.edgeList}
\alias{create.an.edgeList}
\title{Create an edge list}
\usage{
create.an.edgeList(events)
}
\arguments{
\item{events}{dataframe containing all events}
}
\description{
This function will generate an edge list from an events dataframe
}
