% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{create.window.group}
\alias{create.window.group}
\title{Create a window with a group category}
\usage{
create.window.group(df.total, start, end)
}
\arguments{
\item{df.total}{dataframe containing all events}

\item{start}{the starting time of the window}

\item{end}{the ending time of the window}
}
\description{
This function will generate a window from a larger event dataframe, based on the time of the events. This function will keep all other columns, while create.window() will only keep to and from.
}
