% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{create.window.time}
\alias{create.window.time}
\title{Create a window with a time column}
\usage{
create.window.time(df.total, start, end)
}
\arguments{
\item{df.total}{dataframe containing all events}

\item{start}{the starting time of the window}

\item{end}{the ending time of the window}
}
\description{
This function will generate a window from a larger event dataframe, based on the time of the events. The dataframe returned will have to, from, and time as columns.
}
