% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurment_functions_GRAPH.R
\name{degree_mean}
\alias{degree_mean}
\title{Mean degree of a network}
\usage{
degree_mean(net)
}
\arguments{
\item{net}{An igraph network.}
}
\description{
This function will estimate mean degree of a network
}
