% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.scans}
\alias{df.scans}
\title{Data on observational scans of a vervet monkey group}
\format{
An object of class \code{data.frame} with 338 rows and 2 columns.
}
\usage{
df.scans
}
\description{
These data were collected on a group of vervet monkeys in the Samara Game Reserve in South Africa.
}
\keyword{datasets}
