% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurment_functions_DYADS.R
\name{dyad_change}
\alias{dyad_change}
\title{dyad_change function}
\usage{
dyad_change(g1, g2)
}
\arguments{
\item{g1}{First graph.}

\item{g2}{Second graph.}
}
\description{
This function will take two graphs and take the difference in dyad weight.
}
