% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_dyads.R
\name{dyad_diff}
\alias{dyad_diff}
\title{dyad_diff function}
\usage{
dyad_diff(g)
}
\arguments{
\item{g}{graph to extract dyad measures from}
}
\description{
This function will take a graph and take dyad level difference in weights.
}
