% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurment_functions_DYADS.R
\name{dyad_mean}
\alias{dyad_mean}
\title{dyad_mean function}
\usage{
dyad_mean(g)
}
\arguments{
\item{g}{graph to extract dyad measures from}
}
\description{
This function will take a graph and take dyad level mean weight.
}
