% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_dyads.R
\name{dyad_proportion}
\alias{dyad_proportion}
\title{dyad_proportion function}
\usage{
dyad_proportion(g)
}
\arguments{
\item{g}{graph to extract dyad measures from}
}
\description{
This function will take a graph and take dyad proportion of weights.
}
