% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_dyads.R
\name{dyad_sum}
\alias{dyad_sum}
\title{dyad_sum function}
\usage{
dyad_sum(g)
}
\arguments{
\item{g}{graph to extract dyad measures from}
}
\description{
This function will take a graph and take dyad level sum of weights.
}
