% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurment_functions_DYADS.R
\name{dyad_weight}
\alias{dyad_weight}
\title{dyad_weight function}
\usage{
dyad_weight(g)
}
\arguments{
\item{g}{graph to extract dyad measures from}
}
\description{
This function will take a graph and take dyad level sum of weights.
}
