% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{dyadTS.plot}
\alias{dyadTS.plot}
\title{Plot function for dyadTS outputs}
\usage{
dyadTS.plot(data = output.net.dir, legend = FALSE, plotCI = FALSE)
}
\arguments{
\item{data}{Output from the dyadTS function.}

\item{legend}{Whether to include the legend (Default: TRUE)}
}
\description{
This function will plot the output from the dyadTS function
}
