% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_dyads.R
\name{dyadTS}
\alias{dyadTS}
\title{dyadTS function}
\usage{
dyadTS(event.data, windowSize = 30, windowShift = 1, type = "proportion",
  directedNet = T, threshold = 30, windowStart = 0, startDate = NULL,
  order.events = F)
}
\arguments{
\item{event.data}{dataframe containing events between individuals/objects}

\item{windowSize}{size of the window in which to make network measures (should be the same scale as the time column)}

\item{windowShift}{the amount to shift the moving window for each measure}

\item{type}{is the type of measure to take in each window. Currently available are: sum strength, mean strength, proportion weight, relative strength}

\item{directedNet}{Whether the events are directed or no: true or false.}

\item{threshold}{minimum number of events to calculate a network measure (otherwise NA is produced).}

\item{windowStart}{The time of the first window. This should corespond to the first events.}

\item{startDate}{Optional argument to set the date of the first event.}
}
\description{
This function will take a dataframe with events between individuals/objects, and take dyad level measures using a moving window approach.
A time column is required.
}
\examples{

ts.out<-dyadTS(event.data=groomEvents[1:200,])
dyadTS.plot(ts.out)

}
