% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurment_functions_NODES.R
\name{edge.weight.skewness}
\alias{edge.weight.skewness}
\title{Estimate skewness of the edge weight distribution for each node in a graph}
\usage{
edge.weight.skewness(graph1, type = "all")
}
\arguments{
\item{graph1}{An igraph object.}

\item{type}{The type of edges to pull from the graph: 'in', 'out', or 'all'. In non-directed graphs this is ignored.}
}
\description{
This function will calculate the skewness of edge weights from nodes in a graph.
}
\examples{
library(igraph)

#Random graph
graph1 <- random.graph.game(n=10, p.or.m = 0.3)
E(graph1)$weight <- c(1,2,3,4,5,6,7,7,7,7)

#skewness
#edge.weight.skeweness(graph1)

}
