% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurment_functions_EFFORT.R
\name{effort.events}
\alias{effort.events}
\title{Unique sampeling IDs}
\usage{
effort.events(df.window)
}
\arguments{
\item{df.window}{An events data frame with a column (sampleID) containing a unique id for every sampling event (i.e., group scan in a gambit of the group sampling procedure).}
}
\description{
This function will estimate the total sampling, using a count of the time a sample was performed
}
