% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurment_functions_EFFORT.R
\name{effort.focal}
\alias{effort.focal}
\title{Focal sampling}
\usage{
effort.focal(df.window, effortData, directed = FALSE)
}
\arguments{
\item{df.window}{A data frame with focal duration per individual per unique focal, with a date time column.}

\item{effortData}{A data frame with the first column as a datatime stamp, a second column with the ID of the focal, and a thrid column with the total time of the focal.}

\item{directed}{Whether a directed network should be constructed (Default=FALSE).}
}
\description{
This function will estimate the sampling effort from focal data. It will calculate the total durration each dyad had the potential to be observed (i.e., Interactions between A and B could only be seen if A or B was the focal). This sampling correction should be applied post network construction.
}
