% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurment_functions_GRAPH.R
\name{eigen_mean}
\alias{eigen_mean}
\title{Mean eigenvector centrality of a network}
\usage{
eigen_mean(net)
}
\arguments{
\item{net}{An igraph network.}
}
\description{
This function will estimate mean eigenvector centrality of a network
}
