% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{estimate.random.range.perm}
\alias{estimate.random.range.perm}
\title{Permutation test}
\usage{
estimate.random.range.perm(graphW, np, type, directedNet)
}
\arguments{
\item{graphW}{Network (igraph)}

\item{np}{number of permutations}

\item{type}{the measure to use for permutation. Currently available are: betweennes, closeness, eigen, and cc (i.e., clustering coeficient)}

\item{directedNet}{wheter the network is directed or not.}
}
\description{
This function will generate edge permutations for each network
}
