% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{estimate.uncertainty.boot}
\alias{estimate.uncertainty.boot}
\title{Bootstrap estimates}
\usage{
estimate.uncertainty.boot(dataSub, nb, type, directedNet)
}
\arguments{
\item{dataSub}{dataframe containing events}

\item{nb}{number of bootstraps}

\item{type}{the measure to use for the bootstrap. Currently available are: betweennes, closeness, eigen, and cc (i.e., clustering coeficient)}

\item{directedNet}{wheter the network is directed or not.}
}
\description{
This function will generate bootstrapped estimates for each network
}
