% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_graph.R
\name{extract_effort_time}
\alias{extract_effort_time}
\title{Extract effort from an events dataframe}
\usage{
extract_effort_time(data = df.nn, windowsize = days(30),
  windowshift = days(1), effortFun = effort.time)
}
\arguments{
\item{data}{A dataframe with relational data in the first two rows, and a time stamp in the third row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{windowsize}{The size of the moving window in which to take network measures. These should be provided as e.g., days(30), hours(5), ... etc.}

\item{windowshift}{The amount to shift the moving window for each measure. Again times should be provided as e.g., days(1), hours(1), ... etc.}
}
\description{
This function will estimate the amount of sampling time within a window using the min and max sampling times for each day.
}
