% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_graph.R
\name{extract_lagged_measure_network}
\alias{extract_lagged_measure_network}
\title{Extract measures from a list of networks when the measure requires comparisons between networks}
\usage{
extract_lagged_measure_network(netlist, measureFun, lag = 1, firstNet)
}
\arguments{
\item{netlist}{List of networks.}

\item{measureFun}{A function that takes two networks as input and returns a single value. The first network is the lagged network, and the second is the current network.}

\item{lag}{At what lag should networks be compared? The number here will be based on the order of the network list generated. E.g., a list of networks generated using a window shift of 10 days, and a lag of 1, would compare networks 10days apart.}

\item{firstNet}{If TRUE the comparison between networks is always between the current and first network.}
}
\description{
This function will estimate network measures from a list of networks.
}
