% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_dyads.R
\name{extract_measure_dyads}
\alias{extract_measure_dyads}
\title{Extract dyad level network measures from a list of networks}
\usage{
extract_measure_dyads(netlist, measureFun, unique.names)
}
\arguments{
\item{netlist}{List of networks.}

\item{measureFun}{A function that takes a network as input and returns a value for each dyad.}

\item{unique.names}{A list of all dyads/edges in the networks.}
}
\description{
This function will estimate dyad level network measures from a list of networks.
}
\examples{



}
