% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_graph.R
\name{extract_measure_network}
\alias{extract_measure_network}
\title{Extract network measures from a list of networks}
\usage{
extract_measure_network(netlist, measureFun)
}
\arguments{
\item{netlist}{List of networks.}

\item{measureFun}{A function that takes a network as input and returns a single value.}
}
\description{
This function will estimate network measures from a list of networks.
}
