% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_nodes.R
\name{extract_measure_nodes}
\alias{extract_measure_nodes}
\title{Extract node level network measures from a list of networks}
\usage{
extract_measure_nodes(netlist, measureFun, unique.names)
}
\arguments{
\item{netlist}{List of networks.}

\item{measureFun}{A function that takes a network as input and returns a value for each node.}

\item{unique.names}{A list of all names/nodes in the networks.}
}
\description{
This function will estimate node level network measures from a list of networks.
}
\examples{



}
