% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_graph.R
\name{extract_networks_para}
\alias{extract_networks_para}
\title{Extract networks from a moving window using multiple cores}
\usage{
extract_networks_para(data, windowsize, windowshift, directed = FALSE,
  cores = 2, SRI, effortFun = NULL, effortData = NULL)
}
\arguments{
\item{data}{A dataframe with relational data in the first two columns, and a time stamp in the third column. An optional column with a weight can be added if there is a duration or magnitude for each interaction (column name for this should be set to 'weight'). Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{windowsize}{The size of each window in which to generate a network.}

\item{windowshift}{The amount of time to shift the window when generating networks.}

\item{directed}{Whether to consider the network as directed or not (TRUE/FALSE).}

\item{cores}{How many cores should be used.}

\item{SRI}{Wether to use the simple ratio index (Default=FALSE).}

\item{effortFun}{This is a function that takes as input the data within a window of time and returns the total sampling effort.}

\item{effortData}{This is a dataframe containing the data used to calculate sampling effort. The first column should contain timedate values.}
}
\description{
This function will create a time series of networks from a dataframe with relational events and a time stamp, using parallel processing.
}
