% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{get_SRI}
\alias{get_SRI}
\title{Estimate SRI value for each dyad}
\usage{
get_SRI(df.window)
}
\arguments{
\item{graph1}{first graph (igraph)}

\item{graph2}{second graph (igraph)}
}
\description{
This function will calculate the SRI score for each dyad
}
\examples{


}
