% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{graphTS_plot}
\alias{graphTS_plot}
\title{Plot function for graphTS outputs}
\usage{
graphTS_plot(data, plotCI = TRUE)
}
\arguments{
\item{data}{Dataframe with relational data in the first two rows, with weights in the thrid row, and a time stamp in the fourth row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{plotCI}{Whether to plot the confidence intervals (Default: TRUE)}
}
\description{
This function will plot the output from the graphTS function
}
