% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groomEvents}
\alias{groomEvents}
\title{Data on grooming events in a vervet group}
\format{An object of class \code{data.frame} with 3762 rows and 3 columns.}
\usage{
groomEvents
}
\description{
These data were collected on a group of vervet monkeys in the Samara Game Reserve in South Africa.
}
\keyword{datasets}
