% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_graph.R
\name{net.para}
\alias{net.para}
\title{Extract networks in parallel using a dataframe of times}
\usage{
net.para(data, window.ranges, directed = FALSE, effortFun = NULL,
  effortData = NULL)
}
\arguments{
\item{data}{Dataframe with relational data in the first two columns, with weights in the thrid columns, and a time stamp in the fourth row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{window.ranges}{The dataframe containing the start and end times of each window to create a network from.}

\item{directed}{Whether to consider the networks are directed or not.}

\item{effortFun}{This is a function that takes as input the data within a window of time and returns the total sampling effort.}
}
\description{
This function will generate networks in parallel using a dataframe with time constraints.
}
