% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_graph.R
\name{net.window}
\alias{net.window}
\title{Extract one network within time constriants}
\usage{
net.window(data, windowstart, windowend, directed = FALSE)
}
\arguments{
\item{data}{Dataframe with relational data in the first two rows, with weights in the thrid row, and a time stamp in the fourth row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{windowstart}{The start of the window.}

\item{windowend}{The end of the window.}

\item{directed}{Whether to consider the network as directed or not.}
}
\description{
This function will generate one network from a dataframe with time constraints.
}
