% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{netWin}
\alias{netWin}
\title{netWin function}
\usage{
netWin(event.data, nBoot = 100, nPerm = 100, windowSize = 30,
  windowShift = 1, type = "cc", directedNet = T, threshold = 30)
}
\arguments{
\item{event.data}{dataframe containing events between individuals/objects}

\item{nBoot}{number of bootstrap estimates to take for each measure}

\item{nPerm}{number of permutation estimates to take for each measure}

\item{windowSize}{size of the window in which to make network measures (should be the same scale as the time column)}

\item{windowShift}{the amount to shift the moving window for each measure}

\item{type}{is the type of measure to take in each window. Currently available are: betweennes, closeness, eigen, and cc (i.e., clustering coeficient)}

\item{directedNet}{Whether the events are directed or no: true or false.}

\item{threshold}{minimum number of events to calculate a network measure (otherwise NA is produced)}
}
\description{
This function will take a dataframe with events between individuals/objects, and take network measures using a moving window approach.
A time column is required.
}
\examples{

#not run
ts.out<-netTS(event.data=df)
plot.netTS(ts.out)

}
