% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_graph.R
\name{node_first_last}
\alias{node_first_last}
\title{First and last observation time of each node}
\usage{
node_first_last(data)
}
\arguments{
\item{data}{The events dataframe used in the nodeTS function.}
}
\description{
This calculates nodes min and max observed times.
}
