% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_nodes.R
\name{nodeTS.plot}
\alias{nodeTS.plot}
\title{Plotting function for nodeTS dataframes}
\usage{
nodeTS.plot(df.ts, nEvents = FALSE, dates = FALSE)
}
\arguments{
\item{df.ts}{output dataframe from the nodeTS function}

\item{nEvents}{Opional argument to plot the number of events}

\item{dates}{Optional argument to plot the date as opposed to the time since the first event}
}
\description{
This function will plot the output of the nodeTS function
}
\examples{

ts.out<-nodeTS(event.data=groomEvents[1:200,])
nodeTS.plot(ts.out)

}
