% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_dyads.R
\name{order_events}
\alias{order_events}
\title{Order events alphabetically}
\usage{
order_events(event.data)
}
\arguments{
\item{event.data}{dataframe containing events between individuals/objects}
}
\description{
This function will take an event dataframe and order the names alphabetically. Useful in undirected graphs to ensure all edges have the same name through time.
}
