% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{perm.edge.weights}
\alias{perm.edge.weights}
\title{Perform edge weight permutations}
\usage{
perm.edge.weights(data, measureFun, directed = FALSE, nperm = 1000,
  probs = 0.95, SRI = FALSE)
}
\arguments{
\item{data}{Dataframe with relational data in the first two rows, and a time stamp in the third row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{measureFun}{This is a function that takes as an input a igraph network and returns a single value.}

\item{directed}{Whether to consider the network as directed or not (TRUE/FALSE).}

\item{nperm}{Number of permutations to perform before extracting network measures.}

\item{probs}{numeric vector of probabilities with values in [0,1].}
}
\description{
This function will permute a network by randomly switching weights between edges.
}
