% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{perm.events.directed}
\alias{perm.events.directed}
\title{Perform permutation on the events dataframe, maintaining in/out degree of nodes}
\usage{
perm.events.directed(data, measureFun, directed = FALSE, nperm = 1000,
  probs = 0.95, SRI, effort = 1)
}
\arguments{
\item{data}{Dataframe with relational data in the first two rows, with weights in the thrid row, and a time stamp in the fourth row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{measureFun}{This is a function that takes as an input a igraph network and returns a single value.}

\item{directed}{Whether to consider the network as directed or not (TRUE/FALSE).}

\item{nperm}{Number of permutations to perform before extracting network measures.}

\item{probs}{numeric vector of probabilities with values in [0,1].}

\item{SRI}{Wether to use the simple ratio index (Default=FALSE).}

\item{effort}{This is a measure of sampling effort}
}
\description{
This function will permute a network by randomly switching individuals between events, keeping the oringinal order of the interactions (e.g., individuals in the from or to columns are permuted, but no permutations between from and to columns).
}
