% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{perm.events.multiple.outputs}
\alias{perm.events.multiple.outputs}
\title{Perform permutation on the events dataframe}
\usage{
perm.events.multiple.outputs(
  data,
  measureFun,
  directed = FALSE,
  nperm = 1000,
  probs = 0.95,
  SRI = FALSE,
  effortFun = NULL,
  effortData = NULL,
  windowstart,
  windowend
)
}
\arguments{
\item{data}{Dataframe with relational data in the first two rows, and a time stamp in the third row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{measureFun}{This is a function that takes as an input a igraph network and returns a single value.}

\item{directed}{Whether to consider the network as directed or not (TRUE/FALSE).}

\item{nperm}{Number of permutations to perform before extracting network measures.}

\item{probs}{numeric vector of probabilities with values in [0,1].}

\item{SRI}{Wether to use the simple ratio index (Default=FALSE).}

\item{effortFun}{This is a function that takes as input the data within a window of time and returns the total sampling effort.}

\item{effortData}{This is a dataframe containing the data used to calculate sampling effort.}
}
\description{
This function will permute a network by randomly switching individuals between events.
}
