% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{plot.graphTS}
\alias{plot.graphTS}
\title{Plotting function for netTS dataframes}
\usage{
\method{plot}{graphTS}(df.ts)
}
\arguments{
\item{df.ts}{output dataframe from the netTS function}
}
\description{
This function will plot the output of the netTS function
}
\examples{

<<<<<<< HEAD:man/plot.graphTS.Rd
ts.out<-graphTS(event.data=df[1:200,])
plot.graphTS(ts.out)
=======
ts.out<-netWin(event.data=df[1:200,])
plot.netTS(ts.out)
>>>>>>> 2a4ddbbb5fda0fef52e52c81ce2e7c5dc8c6bf45:man/plot.netTS.Rd

}
