% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{plot.netTS}
\alias{plot.netTS}
\title{Plotting function for netTS dataframes}
\usage{
\method{plot}{netTS}(df)
}
\arguments{
\item{df}{output dataframe from the netTS function}
}
\description{
This function will plot the output of the netTS function
}
\examples{

#not run
<<<<<<< HEAD
library(netTS)
ts.out<-netWin(event.data=df)
=======
ts.out<-netTS(event.data=df)
>>>>>>> origin/master
plot.netTS(ts.out)

}
