% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_nodes.R
\name{plot.nodeTS}
\alias{plot.nodeTS}
\title{Plotting function for nodeTS dataframes}
\usage{
\method{plot}{nodeTS}(df.ts)
}
\arguments{
\item{df}{output dataframe from the nodeWin function}
}
\description{
This function will plot the output of the nodeWin function
}
\examples{

ts.out<-nodeWin(event.data=df[1:200,])
plot.nodeTS(ts.out)

}
