% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_dyads.R
\name{trim_dyads}
\alias{trim_dyads}
\title{Trim dyads.}
\usage{
trim_dyads(dyadvalues, data, directed)
}
\arguments{
\item{dyadvalues}{Output from the dyadTS function.}

\item{data}{The events dataframe used in the dyadTS function.}

\item{directed}{Whether to treat the dyads as directed or not}
}
\description{
This function removes dyad values beyond their min and max observed times.
}
