% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_graph.R
\name{trim_graph}
\alias{trim_graph}
\title{Trim nodes when taking network measures.}
\usage{
trim_graph(df.window, data, windowstart, windowend)
}
\arguments{
\item{data}{The events dataframe used in the nodeTS function.}

\item{nodevalues}{Output from the nodeTS function.}
}
\description{
This function removes node from the network when they are beyond their min and max observed times, then takes the network measure.
}
