% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_nodes.R
\name{trim_nodes}
\alias{trim_nodes}
\title{Trim nodes.}
\usage{
trim_nodes(nodevalues, data)
}
\arguments{
\item{nodevalues}{Output from the nodeTS function.}

\item{data}{The events dataframe used in the nodeTS function.}
}
\description{
This function removes node values beyond their min and max observed times.
}
