% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_graph.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Calculate the weighted average correcting for nodes enter or leaving the network}
\usage{
weighted_mean(obs.values, inOut, net)
}
\arguments{
\item{obs.values}{Vector of node level network measures, with column names for each individual measure.}

\item{inOut}{A dataframe with names in the first column, in a second column the dates these individuals first entered the network, and in a third column the dates these individuals left the network.}

\item{net}{The network used to calculate the weighted mean.}
}
\description{
This function will estimate the weights of node based on their first and last observation and the time within a given window (e.g., due to births or deaths), and calculate the weighted average network value.
}
