% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_graph.R
\name{weightedAvg}
\alias{weightedAvg}
\title{Calculate the weighted average of network measures}
\usage{
weightedAvg(obs.values, inOut, windowstart, windowend)
}
\arguments{
\item{obs.values}{Vector of node level network measures, with column names for each individual measure.}

\item{inOut}{A dataframe with names in the first column, in a second column the dates these individuals first entered the network, and in a third column the dates these individuals left the network.}

\item{windowstart}{The starting date of the window.}

\item{windowend}{The ending date of the window.}
}
\description{
This function will estimate the weighted average of node level measures based on their leaving and entering each network (e.g., due to births or deaths).
}
