% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{window_check_plot}
\alias{window_check_plot}
\title{Plot function for windowsize check outputs}
\usage{
window_check_plot(data, legend = TRUE)
}
\arguments{
\item{data}{Output from the window.check function.}

\item{legend}{Whether to include the legend (Default: TRUE)}
}
\description{
This function will plot the output from the window.check function
}
