% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{windowsize.check}
\alias{windowsize.check}
\title{Bootstrap convergence check with subsampling}
\usage{
windowsize.check(data, windowsize = days(30), windowshift = days(1),
  directed = FALSE, measureFun = degree, corFun = 1, boot.samples = 100,
  SRI = FALSE, probs = c(0.025, 0.975), subsamples = c(1, 0.8, 0.6),
  plot = TRUE)
}
\arguments{
\item{data}{Dataframe with relational data in the first two rows, and a time stamp in the third row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{windowsize}{The size of each window in which to generate a network.}

\item{windowshift}{The amount of time to shift the window when generating networks.}

\item{directed}{Whether to consider the network as directed or not (TRUE/FALSE).}

\item{measureFun}{The measurment function to perform the bootstap on (should be at the node level).}

\item{corFun}{The method used to compare observed node/dyad values with bootstrapped values: 1-Cosine similarity, 2-pearsons correlation}

\item{boot.samples}{The number of bootstrapped samples to run (Default=100)}

\item{SRI}{Wether to use the simple ratio index (Default=FALSE)}

\item{probs}{The quantiles of the bootrap samples to return (Default=c(0.025,0.975)).}

\item{subsamples}{A vector of values between 0-1 used to subsample the original dataframe.}

\item{plot}{Wether a plot of the results should be produced.}
}
\description{
This function will estimate the convergence of the chosen network measure using bootstrapped samples of the data, while also checking for measurement sensitivity to data subsampling.
}
