% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{cosine_between_graphs}
\alias{cosine_between_graphs}
\title{Estimate consine similarity between graphs}
\usage{
cosine_between_graphs(graph1, graph2)
}
\arguments{
\item{graph1}{first graph (igraph)}

\item{graph2}{second graph (igraph)}
}
\description{
This function will calculate the cosine similarity between two graphs
}
\examples{

#two random graphs
<<<<<<< HEAD
library(igraph)
=======
>>>>>>> f0dd017bcfa4ef02c08b3dcee5f0d9b349b0fcdc
graph1 <-random.graph.game(n=10,p=0.1)
graph2 <-random.graph.game(n=10,p=0.1)
cosine_between_graphs(graph1,graph2)

#moving windo with cosine similarity
ts.out<-netWin(event.data=df[1:100,], type='cosine')
plot.netTS(ts.out)

}
