% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{cosine_between_graphs}
\alias{cosine_between_graphs}
\title{Estimate consine similarity between graphs}
\usage{
cosine_between_graphs(graph1, graph2)
}
\arguments{
\item{graph1}{first graph (igraph)}

\item{graph2}{second graph (igraph)}
}
\description{
This function will calculate the cosine similarity between two graphs
}
\examples{

#two random graphs
graph1 <-random.graph.game(n=10,p=0.1)
graph2 <-random.graph.game(n=10,p=0.1)
cosine_between_graphs(graph1,graph2)

}
