% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_nodes.R
\name{cosine_between_graphs_nodes}
\alias{cosine_between_graphs_nodes}
\title{Estimate consine similarity between nodes in a graph}
\usage{
cosine_between_graphs_nodes(graph1, graph2)
}
\arguments{
\item{graph1}{first graph (igraph)}

\item{graph2}{second graph (igraph)}
}
\description{
This function will calculate the cosine similarity between nodes at two time periods.
}
\examples{

#two random graphs
library(igraph)
graph1 <-erdos.renyi.game(n=10,p=0.1)
graph2 <-erdos.renyi.game(n=10,p=0.1)
cosine_between_graphs_nodes(graph1,graph2)

#moving windo with cosine similarity
ts.out<-nodeTS(event.data=groomEvents[1:100,], type='cosine')
nodeTS.plot(ts.out)

}
