% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{create.a.network}
\alias{create.a.network}
\title{Create a network}
\usage{
create.a.network(data, directed = FALSE, SRI = FALSE)
}
\arguments{
\item{data}{Dataframe containing all events. The first two coloums should contain the 'to' and 'from' indentities involved in the interaction, while the third column should contain the 'weight' of the interaction.}

\item{directed}{Treat the network as directed or not (Default=FALSE).}

\item{SRI}{Wether to use the simple ratio index or not (Default=FALSE).}
}
\description{
This function will generate a network from an events dataframe.
}
