% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{create.a.network}
\alias{create.a.network}
\title{Create a network from an edge list}
\usage{
create.a.network(events, directedG = TRUE)
}
\arguments{
\item{events}{dataframe containing all events.}

\item{directedG}{Treat the network as directed or not (Default = TRUE)}
}
\description{
This function will generate a network from an event dataframe
}
