% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{create.an.edgeList}
\alias{create.an.edgeList}
\title{Create an edge list}
\usage{
create.an.edgeList(data, effort = 1)
}
\arguments{
\item{data}{Dataframe containing all events. The first two coloums should contain the 'to' and 'from' indentities involved in the interaction, while the third column should contain the 'weight' of the interaction.}

\item{effort}{The number of scans, or measure of sampling effort (e.g., hours sampling). The default of 1 assumes equal sampling effort througout.}
}
\description{
This function will generate an edge list from an events dataframe
}
