% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{create.window}
\alias{create.window}
\title{Create a window}
\usage{
create.window(data, start, end)
}
\arguments{
\item{data}{dataframe containing all events}

\item{start}{the starting time of the window}

\item{end}{the ending time of the window}
}
\description{
This function will generate a window from a larger event dataframe, based on the time of the events
}
