% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_dyads.R
\name{dyadTS.plot}
\alias{dyadTS.plot}
\title{Plotting function for dyadTS dataframes}
\usage{
dyadTS.plot(df.ts, nEvents = FALSE, dates = FALSE)
}
\arguments{
\item{df.ts}{output dataframe from the dyadTS function}

\item{nEvents}{Opional argument to plot the number of events}

\item{dates}{Optional argument to plot the date as opposed to the time since the first event}
}
\description{
This function will plot the output of the dyadTS function
}
\examples{

ts.out<-dyadTS(event.data=groomEvents[1:200,])
dyadTS.plot(ts.out)

}
