% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurment_functions_EFFORT.R
\name{effort.scan}
\alias{effort.scan}
\title{Unique sampling IDs}
\usage{
effort.scan(df.window, directed = FALSE)
}
\arguments{
\item{df.window}{The events data frame with interactions.}

\item{directed}{Whether a directed network should be constructed (Default=FALSE).}
}
\description{
This function will estimate the total sampling, using a count of the time a sample was performed. It will return a network with edges corrected by the number of scans.
}
