% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurment_functions_EFFORT.R
\name{effort.scan}
\alias{effort.scan}
\title{Unique scan IDs}
\usage{
effort.scan(df.window, df.scans, directed = FALSE)
}
\arguments{
\item{df.window}{The events data frame with interactions.}

\item{df.scans}{A dataframe with the first column as a data column, and a second column with the number of scans durring that day.}

\item{directed}{Whether a directed network should be constructed (Default=FALSE).}
}
\description{
This function will correct edge weights in a network, using a count of the number of scans used to construct the network.
}
