% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{estimate.random.range.perm}
\alias{estimate.random.range.perm}
\title{Permutation test}
\usage{
estimate.random.range.perm(graphW, np, type, directedNet, previousNet = NULL)
}
\arguments{
\item{graphW}{Network (igraph)}

\item{np}{number of permutations}

\item{type}{the measure to use for permutation. Currently available are: between (i.e., mean betweenness centrality),
close (i.e., mean closenes), eigen (i.e., eigen vector centrality), and cc (i.e., clustering coeficient). See the igraph package
for details on each measure.}

\item{directedNet}{wheter the network is directed or not.}

\item{previousNet}{A second network to compare disimilariy (e.g., for cosine similariy).}
}
\description{
This function will generate edge permutations for each network
}
