% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{graphTS.plot}
\alias{graphTS.plot}
\title{Plot function for graphTS outputs}
\usage{
graphTS.plot(data, plotCI = FALSE)
}
\arguments{
\item{data}{Output from the graphTS function.}

\item{plotCI}{Whether to plot the confidence intervals (Default: TRUE)}
}
\description{
This function will plot the output from the graphTS function
}
