% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{graphTS}
\alias{graphTS}
\title{graphTS function}
\usage{
graphTS(event.data, nBoot = 100, nPerm = 100, windowSize = 30,
  windowShift = 1, type = "cc", directedNet = T, threshold = 30,
  windowStart = 0, lag = 1)
}
\arguments{
\item{event.data}{dataframe containing events between individuals/objects. This dataframe should have a 'to' and a 'from' column, as well as 'time' column representing the time between the start of observations and the events.}

\item{nBoot}{number of bootstrap estimates to take for each measure}

\item{nPerm}{number of permutation estimates to take for each measure}

\item{windowSize}{size of the window in which to make network measures (should be the same scale as the time column)}

\item{windowShift}{the amount to shift the moving window for each measure}

\item{type}{is the type of measure to take in each window. Currently available are: between (i.e., mean betweenness centrality),
close (i.e., mean closenes), eigen (i.e., eigen vector centrality), and cc (i.e., clustering coeficient). See the igraph package
for details on each measure.}

\item{directedNet}{Whether the events are directed or no: true or false.}

\item{threshold}{minimum number of events to calculate a network measure (otherwise NA is produced)}

\item{windowStart}{The time of the first window. This should corespond to the first events.}

\item{lag}{The lag at which to calculate cosine similarity in network structure.}
}
\description{
This function will take a dataframe with events between individuals/objects, and take network measures using a moving window approach.
}
\examples{

ts.out<-graphTS(event.data=groomEvents[1:200,])
graphTS.plot(ts.out)

}
