% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groomEvents}
\alias{groomEvents}
\title{Data on grooming events in a baboon troop}
\format{A dataframe with 837 rows and 3 variables
\describe{
  \item{from}{the identity of the individual doing the grooming}
  \item{to}{the identity of the individual being groomed}
  \item{time}{the day on which the grooming took place, measured as days since observations started}
  ...
}}
\source{
\href{https://www.researchgate.net/profile/Peter_Henzi/publication/312584929_The_strength_of_weak_ties_among_female_baboons_fitness-related_benefits_of_social_bonds/links/5883456892851c21ff445b63/The-strength-of-weak-ties-among-female-baboons-fitness-related-benefits-of-social-bonds.pdf}{download}
}
\usage{
groomEvents
}
\description{
These data were collected on a troop of baboons in the DeHoop Nature reserve in South Africa.
}
\references{
McFarland et al. (2017) Animal Behaviour 126:101-106
(\href{http://www.sciencedirect.com/science/article/pii/S0003347217300362}{sciencedirect})
}
\keyword{datasets}
