% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{nodeTS.plot}
\alias{nodeTS.plot}
\title{Plot function for nodeTS outputs}
\usage{
nodeTS.plot(data, legend = TRUE, plotCI = FALSE)
}
\arguments{
\item{data}{Output from the nodeTS function.}

\item{legend}{Whether to include the legend (Default: TRUE)}
}
\description{
This function will plot the output from the nodeTS function
}
