% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_nodes.R
\name{nodeTS.plot}
\alias{nodeTS.plot}
\title{Plotting function for nodeTS dataframes}
\usage{
nodeTS.plot(df.ts)
}
\arguments{
\item{df.ts}{output dataframe from the nodeWin function}
}
\description{
This function will plot the output of the nodeWin function
}
\examples{

ts.out<-nodeTS(event.data=groomEvents[1:200,])
nodeTS.plot(ts.out)

}
