% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_nodes.R
\name{nodeTS}
\alias{nodeTS}
\title{nodeTS function}
\usage{
nodeTS(
  data,
  windowsize = days(30),
  windowshift = days(1),
  measureFun = degree,
  effortFun = NULL,
  effortData = NULL,
  directed = FALSE,
  lagged = FALSE,
  lag = 1,
  firstNet = FALSE,
  cores = 1,
  permutationFun = perm.events.multiple.outputs,
  nperm = 0,
  probs = 0.95,
  SRI = FALSE
)
}
\arguments{
\item{data}{A dataframe with relational data in the first two columns, and a time stamp in the third columns Note: time stamps should be in ymd_hms format. The lubridate package can be very helpful in organizing times. Note: names in the first two columns are case sensitive.}

\item{windowsize}{The size of the moving window in which to take network measures. These should be provided as e.g., days(30), hours(5), ... etc.}

\item{windowshift}{The amount to shift the moving window for each measure. Again times should be provided as e.g., days(1), hours(1), ... etc.}

\item{measureFun}{This is a function that takes as an input a igraph network and returns values for each node in the network. There are functions within netTS (see details), and custom made functions can be used.}

\item{effortFun}{This is a function that takes as input the data within a window of time and returns the total sampling effort.}

\item{effortData}{This is a dataframe containing the data used to calculate sampling effort.}

\item{directed}{Whether the events are directed or no: true or false.}

\item{lagged}{Whether the network measure function used requires the comparison between two networks. e.g., comparing the current network to one lagged by 10 days. If TRUE the measureFun should take two graphs as input and return a single value. The order of inputs in the function is the lagged network followed by the current network.}

\item{lag}{If lagged is set to TRUE, this is the lag at which to compare networks.}

\item{firstNet}{If lagged is set to TRUE, this compares the subsequent networks to the first network.}

\item{cores}{This allows for multiple cores to be used while generating networks and calculating network measures.}

\item{permutationFun}{This is a function that takes as input an events dataframe and a measurment function, and will return a quantile range of network values (see permutation vignette).}

\item{nperm}{This allows for the estimation the network measure assuming random permutations. Currently the 95 percent quantiles are returned.}

\item{probs}{When nperm > 0 this will determine the probability of the permutation values returned from the permuations.}

\item{SRI}{Wether to use the simple ratio index (Default=FALSE).}
}
\description{
This function will take a dataframe with events between individuals/objects, and take network measures using a moving window approach.
}
\examples{

ts.out<-nodeTS(data=groomEvents)

}
