% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_graph.R
\name{permutation.graph.values}
\alias{permutation.graph.values}
\title{Use permutation to extract uncertainty}
\usage{
permutation.graph.values(data, windowsize, windowshift, directed = FALSE,
  measureFun, probs = 0.95, nperm = 1000)
}
\arguments{
\item{data}{Dataframe with relational data in the first two rows, with weights in the thrid row, and a time stamp in the fourth row. Note: time stamps should be in ymd or ymd_hms format. The lubridate package can be very helpful in organizing times.}

\item{windowsize}{The size of each window in which to generate a network.}

\item{windowshift}{The amount of time to shift the window when generating networks.}

\item{directed}{Whether to consider the network as directed or not (TRUE/FALSE).}

\item{measureFun}{This is a function that takes as an input a igraph network and returns a single value.}

\item{probs}{numeric vector of probabilities with values in [0,1].}

\item{nperm}{Number of permutations to perform before extracting network measures.}
}
\description{
This function will estimate network measures given random permutations on the original data.
}
\examples{



}
