% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions_dyads.R
\name{re.net}
\alias{re.net}
\title{re.net function}
\usage{
re.net(df.int, summary = FALSE)
}
\arguments{
\item{df.int}{Dataframe of the random effects where each row is a sample for each dyadic relationship. The column names of such a dataframe should have both individual names seperated by a "_".}

\item{summary}{If summary is F (Default) then a list of networks is returned, one for each sample. If summary is T then one network is returned with mean edge weights.}
}
\description{
This function will take the random effect matrix from a mixed model and create networks.
}
