% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim.events.data}
\alias{sim.events.data}
\title{Simulate some observation data}
\usage{
sim.events.data(nodes, sampling.periods, sampling.periods.per.day = 1,
  true.net = NULL, A = NULL, B = NULL, C = NULL, D = NULL, E = NULL,
  covariates = NULL)
}
\arguments{
\item{nodes}{Number of individuals within a simulated group.}

\item{sampling.periods}{The number of times the simulated group is observed.}

\item{sampling.periods.per.day}{The number of sampling perids per day.}

\item{true.net}{(Optional) A true underlying network describing the probability of each individual interacting.}

\item{A}{(Optional) A vector specifying the average probability of observing the individual performing the behaivour. Should be the same length as the number of nodes. Default: runif(node,0,1)}

\item{B}{(Optional) A matrix (node x node) specifying the interdependence of individual mean probabilities. Default: matrix of zeros}

\item{C}{(Optional) A matrix specifying the effects of each covariate value on the mean probability. Default: matrix of zeros}

\item{D}{(Optional) A matrix specifying the interdependence of individual variations probabilities. Default: matrix of zeros}

\item{E}{(Optional) A value for the standard deviation around observed probability of observing a behaviour. Default: 0.1}

\item{covariates}{A matrix with covariate values for each sampling period. Should be of size (sampleing.periods x number of covariates)}
}
\description{
This function will simulate events data.
}
