% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim.events.data}
\alias{sim.events.data}
\title{Simulate some observation data}
\usage{
sim.events.data(nodes, sampling.periods, sampling.periods.per.day = 1,
  true.net = NULL, A = NULL, B = NULL, C = NULL, D = NULL, E = NULL,
  covariates = NULL)
}
\arguments{
\item{nodes}{Number of individuals within a simulated group.}

\item{sampling.periods}{The number of times the simulated group is observed.}

\item{sampling.periods.per.day}{The number of sampling periods per day.}

\item{true.net}{(Optional) A network describing who can interact with whom.}

\item{A}{(Optional) A vector specifying the average probability interaction for each individual. Should be the same length as the number of nodes. Default: runif(node,0,1)}

\item{B}{(Optional) A matrix (node x node) specifying the inter-dependence of individual interaction probabilities. Default: matrix of zeros}

\item{C}{(Optional) A matrix (node X covariates) specifying the effect of covariate values on each individual's interaction probability. Default: matrix of zeros}

\item{D}{(Optional) A matrix (node x node) specifying the inter-dependence of noise around individual variations probabilities. Default: matrix of zeros}

\item{E}{(Optional) A value for the standard deviation (noise) around individual's mean interaction probability. Default: 0.1}

\item{covariates}{A matrix (sampling period x covariates) with covariate values for each sampling period.}
}
\description{
This function simulates interaction data based on a group scaning methodology. For a given number of sampling periods, the simulation will generate interactions based on a user defined network, and individual interaction probabilities.
}
