% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim.events.data}
\alias{sim.events.data}
\title{Simulate some observation data}
\usage{
sim.events.data(nodes, sampling.periods, sampling.periods.per.day = 1,
  true.net = NULL, ind.probs = NULL, cor.mat = NULL, ind.sd = NULL)
}
\arguments{
\item{nodes}{Number of individuals within a simulated group.}

\item{sampling.periods}{The number of times the simulated group is observed.}

\item{sampling.periods.per.day}{The number of sampling perids per day.}

\item{true.net}{(Optional) A true underlying network describing the probability of each individual interacting.}
}
\description{
This function will simulate events data.
}
